refactor: 클래스 이름 변경 및 필요없는 클래스 삭제
