+++ src/main/java/chess/controller/SparkChessController.java
@@ -17,7 +17,7 @@ import java.sql.SQLException;
 import java.util.Collections;
 import java.util.List;
 
-public class WebUIChessController {
+public class SparkChessController {
     public static final Gson gson = new Gson();
 
     private final RoomDAO roomDAO = new RoomDAO();
